% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/burdenFirth.R
\name{burdenFirth}
\alias{burdenFirth}
\title{Run the burden Firth test}
\usage{
burdenFirth(formula, data, Z, snpinfo, aggregateBy = "gene",
  snpNames = "Name", mafRange = c(0, 0.5))
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.The details
of model specification are given under 'Details' in the "lm" help file.}

\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{Z}{A genotype matrix (dosage matrix) - rows correspond to
individuals and columns correspond to SNPs. Use 'NA' for missing values.
The row names of this matrix should correspond to the subject names in the
phenotype file.  The column names of this matrix should correspond to SNP
names in the SNP information file.}

\item{snpinfo}{SNP Info data.frame. Must contain fields given in 'snpName'
and 'aggregateBy'}

\item{aggregateBy}{The column in snpinfo on which the individual snps are
to be aggregated by. Default is 'gene'.}

\item{snpNames}{The column in snpinfo where the SNP identifiers.
Defualt is 'Name'}

\item{mafRange}{Range of MAF's to include in the analysis (endpoints
included). Default is all SNPs (0 <= MAF <= 0.5).}
}
\value{
Results of the burden Firth test.
}
\description{
This function runs the burden Firth test.
}
\details{
This function is a wrapper for the logistf function in the logistf
package used in the Firth test.
}

