% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codeToMinor.R
\name{codeToMinor}
\alias{codeToMinor}
\title{Code to the Minor Allele}
\usage{
codeToMinor(Z)
}
\arguments{
\item{Z}{A genotype matrix (dosage matrix) - rows correspond to
individuals and columns correspond to SNPs. Use 'NA' for missing values.}
}
\value{
a recoded genotype matrix (dosage matrix)
}
\description{
This function ensures the genotype is coded to the minor allele.
}
\details{
This function ensures the minor allele frequency is < 0.5.  Note
this assumes the genotype matrix is coded 0/1/2.

If Z is not a matrix it will be coerced with as.matrix.
}

