% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeToMean.R
\name{imputeToMean}
\alias{imputeToMean}
\title{Impute missing values to the mean dosage}
\usage{
imputeToMean(Z)
}
\arguments{
\item{Z}{A genotype matrix (dosage matrix) - rows correspond to
individuals and columns correspond to SNPs. Use 'NA' for missing values.}
}
\value{
a genotype matrix (dosage matrix)
}
\description{
This function imputes missing genotype values to the mean
dosage.
}
\details{
This function imputes missing genotypes to the SNP mean allele
frequency.

If Z is not a matrix it will be coerced with as.matrix.
}

