% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singlesnpFirth.R
\name{singlesnpFirth}
\alias{singlesnpFirth}
\title{Run the single snp Firth test}
\usage{
singlesnpFirth(formula, data, Z, snpinfo = NULL, snpNames = NULL,
  refCol = NULL, altCol = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.The details
of model specification are given under 'Details' in the "lm" help file.}

\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{Z}{A genotype matrix (dosage matrix) - rows correspond to
individuals and columns correspond to SNPs. Use 'NA' for missing values.
The row names of this matrix should correspond to the subject names in the
phenotype file.  The column names of this matrix should correspond to SNP
names in the SNP information file.}

\item{snpinfo}{(optional) SNP Info data.frame.  See Details.}

\item{snpNames}{(optional) The column in snpinfo where the SNP identifiers
are found. See Details.}

\item{refCol}{(optional) The column in snpinfo where the reference allele
for the SNP identifier is found.  See Details.}

\item{altCol}{(optional) The column in snpinfo where the alternate allele
for the SNP identifier is found.  See Details.}
}
\value{
Results of the Firth test in a RAREMETAL friendly format.
}
\description{
This function runs the single snp Firth test.
}
\details{
This function is a wrapper for the logistf function in the logistf
package used in the Firth test. \code{\link[logistf]{logistf}}
}

